+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Reject RSA key generation shorter than 512 bits
 * Fix incorrect exception when SunJSSE validates RSA signatures backed by ACCP RSA
 * Make the provider actually serializable to keep JTREG happy
+* Moved property and resource access to inside PrivilegedAction blocks
 
 ## 1.0.4
 ### Maintenance