+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 1.1.0
 
+### Improvements
+* Now supports DH key agreement for more than two parties.
+
 ### Patches
 * Reject RSA key generation shorter than 512 bits
 * Fix incorrect exception when SunJSSE validates RSA signatures backed by ACCP RSA